Refactoring: Simplifying class hierarchy for bulk export authentication methods.
